Nothing older left is not nothing to read
Two failures the integration lane found, both mine.

A refused reading writes '-infinity' into last_activity_at — the row
exists to hold the refusal count and no read has happened — and scanning
that into a *time.Time failed the whole pass. It is read through NULLIF
now, because "never read" is exactly what a thread with no row answers.

And the window sent a thread backwards when nothing was newer than the
last read, which is not the same as nothing to read: a conversation
grows at the SAME instant as its newest message — the count moves and
the clock does not, which is why the scan records a count at all — and a
thread whose history is fully covered has an empty older window every
pass. Both read nothing, so the model was never called, which is this
ticket's own defect reintroduced at the other end. An empty older window
falls back to the newest one.

The cursor is set after that fallback rather than inside the read, so a
window that was discarded cannot move the reading backwards past text
nobody saw.

The window logic moved to its own file on the way: the read crossed the
500-line cap, and the two halves answer different questions — which
conversations are due, and which part of one to look at.
